created comments.ts and move comment logick there
